Bumped to 2.9.5
Fixed an issue where StatusCode was not included in FromOutcome. Changed this so that StatusCode is now backed by the Keys dictionary. Since StatusCode is a property, it should still be serialized the same way, so this is not a breaking  change for the edge case where an Outcome is serialized to Json and StatusCode is expected.